Megacommit breaks out fields from models
