+++ barbados/barbadosdb/models.py
@@ -1,59 +1,10 @@
 from django.contrib.auth import models as auth_models
-from django.core.exceptions import ValidationError
 from django.utils.translation import ugettext_lazy as _
 from django.db import models
 
+from . import fields
+
 import uuid
-import re
-from datetime import date
-
-
-class PhoneNumberField(models.CharField):
-    """Add rules about phone numbers here
-    """
-
-    def validate(self, value, instance):
-        """Check the value
-        """
-
-        super().validate(value, instance)
-
-        if not value.startswith('+'):
-            raise ValidationError(_('Use international format with phone numbers'))
-
-
-class RegistrationNumberField(models.CharField):
-    """Add rules about registration numbers here.
-    Registration numbers are /[A-Z][0-9]{1-5}/, e.g. H54321.
-    People often write them with a space or dash, e.g. H 54321 or H-54321,
-    but we will require them in the canonical form.
-    """
-
-    def validate(self, value, instance):
-        """Check the value
-        """
-
-        super().validate(value, instance)
-
-        if len(value) > 0 and not re.match('[A-Z][0-9]{1,5}', value):
-            raise ValidationError(_(
-                'Registration number must be blank or of form A12345'))
-
-
-class InspectionYearField(models.IntegerField):
-    """Add rules about inspection years here.
-    """
-
-    def validate(self, value, instance):
-        """Check the value
-        """
-
-        super().validate(value, instance)
-
-        if value is not None and \
-                (value < 1000 or value > date.today().year):
-            raise ValidationError(_(
-                'Inspection date must be 4-digit year and not in future'))
 
 
 # Create your models here.
@@ -68,7 +19,7 @@ class User(auth_models.AbstractUser):
 
     birth_date = models.DateField(null=True, blank=True, default=None)
 
-    phone_number = PhoneNumberField(max_length=24, blank=True, default='')
+    phone_number = fields.PhoneNumberField(max_length=24, blank=True, default='')
 
     # Make this pretty free-form for compatibility with different countries
     street_address = models.TextField(null=True, blank=True, default=None)
@@ -121,7 +72,7 @@ class Boat(models.Model):
 
     manufacturer = models.CharField(max_length=64, blank=True, default='')
 
-    registration_number = RegistrationNumberField(
+    registration_number = fields.RegistrationNumberField(
         max_length=6, blank=True, default='')
 
     sail_number = models.IntegerField(null=True, blank=True, default=None)
@@ -154,9 +105,9 @@ class Boat(models.Model):
         default=None
     )
 
-    inspection_year = InspectionYearField(null=True, blank=True, default=None)
+    inspection_year = fields.InspectionYearField(null=True, blank=True, default=None)
 
-    hull_inspection_year = InspectionYearField(
+    hull_inspection_year = fields.InspectionYearField(
         null=True, blank=True, default=None)
 
     insurance_company = models.CharField(max_length=64, blank=True, default='')
