+++ barbados/barbadosdb/tests.py
@@ -1,4 +1,4 @@
-from . import models
+from . import fields, models
 
 import pytest
 import factory
@@ -79,7 +79,7 @@ def test_bad_phone_number():
         u.full_clean()
 
         pytest.fail(msg='User should not be valid')
-    except models.ValidationError:
+    except fields.ValidationError:
         pass
 
 
@@ -139,7 +139,7 @@ def test_bad_boat_type():
         b.full_clean()
 
         pytest.fail(msg='Boat should not be valid')
-    except models.ValidationError:
+    except fields.ValidationError:
         pass
 
 
@@ -150,7 +150,7 @@ def test_bad_registration_number():
         b.full_clean()
 
         pytest.fail(msg='Boat should not be valid')
-    except models.ValidationError:
+    except fields.ValidationError:
         pass
 
 
@@ -161,7 +161,7 @@ def test_bad_material():
         b.full_clean()
 
         pytest.fail(msg='Boat should not be valid')
-    except models.ValidationError:
+    except fields.ValidationError:
         pass
 
 
@@ -172,7 +172,7 @@ def test_bad_inspection_class():
         b.full_clean()
 
         pytest.fail(msg='Boat should not be valid')
-    except models.ValidationError:
+    except fields.ValidationError:
         pass
 
 
@@ -183,7 +183,7 @@ def test_bad_inspection_year_too_short():
         b.full_clean()
 
         pytest.fail(msg='Boat should not be valid')
-    except models.ValidationError:
+    except fields.ValidationError:
         pass
 
 
@@ -194,7 +194,7 @@ def test_bad_inspection_year_in_future():
         b.full_clean()
 
         pytest.fail(msg='Boat should not be valid')
-    except models.ValidationError:
+    except fields.ValidationError:
         pass
 
 
@@ -205,7 +205,7 @@ def test_bad_hull_inspection_year_too_short():
         b.full_clean()
 
         pytest.fail(msg='Boat should not be valid')
-    except models.ValidationError:
+    except fields.ValidationError:
         pass
 
 
@@ -216,7 +216,7 @@ def test_bad_hull_inspection_year_in_future():
         b.full_clean()
 
         pytest.fail(msg='Boat should not be valid')
-    except models.ValidationError:
+    except fields.ValidationError:
         pass
